backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy>=2.0
aiosqlite>=0.20
pydantic[email]>=2.11
python-multipart>=0.0.18
openpyxl>=3.1
aiosmtplib>=3.0
email-validator>=2.2
python-dotenv>=1.0
jinja2>=3.1
